Modified scale manager to fix bug
When changing preset, during the replaceState method for the APVTS, the order of the parameters meant that the grid cell parameters are set before scale type and base note. This meant that during the method call the grid cell parameter callbacks were being called before the scale type and base note params were set (even though they weren't being set to a new value). This meant that when the cell callbacks referenced the scale type and base note params directly from the APVTS, they always gave the default value as if the grid was in chromatic mode. I've fixed this now by changing the scale manager to save the current scale and base note values in separate variables and only update them when the apvts param is updated.